1.免费试吃基本完成

+++ plant-core/plant-entity/src/main/java/com/moguying/plant/core/entity/taste/Taste.java
@@ -35,7 +35,7 @@ public class Taste {
      * 试吃份数
      */
     @JSONField(ordinal = 5)
-    private Integer tasteCount;
+    private Long tasteCount;
 
 
     /**
